requirements.txt
pandas
pyodbc
mysqlclient
mysql-connector-python

